退货小数；修改密码

+++ app/src/main/java/com/runwise/supply/entity/TransferDetailResponse.java
@@ -112,7 +112,7 @@ public class TransferDetailResponse {
             this.priceUnit = priceUnit;
         }
 
-        public void setActualOutputNum(int actualOutputNum) {
+        public void setActualOutputNum(double actualOutputNum) {
             this.actualOutputNum = actualOutputNum;
         }
